--- a/src/ProgressOnderwijsUtils/Pocos/BulkInsertTarget.cs
+++ b/src/ProgressOnderwijsUtils/Pocos/BulkInsertTarget.cs
@@ -60,6 +60,6 @@ namespace ProgressOnderwijsUtils
         }
 
         public void BulkInsert(SqlConnection sqlConn, DbDataReader dbDataReader, string sourceNameForTracing, CommandTimeout timeout = default)
-            => BulkInsertImplementation.Execute(sqlConn, TableName, Columns, Mode, Options, timeout, dbDataReader, sourceNameForTracing);
+            => BulkInsertImplementation.Execute(sqlConn, dbDataReader, this, sourceNameForTracing, timeout);
     }
 }