… ProgressOnderwijsUtils.sln.DotSettings
… test/ProgressOnderwijsUtils.Tests/Data/BulkInsertTest.cs
… test/ProgressOnderwijsUtils.Tests/Data/PocoBulkCopyTest.cs
… tools/ProgressOnderwijsUtilsBenchmarks/SmallBatchInsertBench.cs