Use plain insert statements for small batches.


Baseline (24c7dbfe1147141d8fe0f21ce6b25295b9268be6): 



<pre><code>
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.508 (2004/?/20H1)
Intel Core i7-8700K CPU 3.70GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.100-rc.1.20452.10
  [Host] : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT
</code></pre>
<pre><code>Job=InProcess  MaxRelativeError=0.002  OutlierMode=DontRemove  
Toolchain=InProcessNoEmitToolchain  IterationCount=15  LaunchCount=1  
MaxIterationCount=5000  WarmupCount=3  
</code></pre>

<table>
<thead><tr><th>                          Method</th><th>  Mean</th><th>Error</th><th>StdDev</th><th>Gen 0</th><th>Gen 1</th><th>Gen 2</th><th>Allocated</th>
</tr>
</thead><tbody><tr><td>BulkInsertTarget_LoadFromTable</td><td>89.511 μs</td><td>0.8805 μs</td><td>0.8236 μs</td><td>1.0986</td><td>-</td><td>-</td><td>7.46 KB</td>
</tr><tr><td>BulkCopyToSqlServer_cachedTarget_0</td><td>2.681 μs</td><td>0.0208 μs</td><td>0.0194 μs</td><td>0.4845</td><td>-</td><td>-</td><td>2.98 KB</td>
</tr><tr><td>BulkCopyToSqlServer_cachedTarget_1</td><td>1,220.698 μs</td><td>8.6026 μs</td><td>8.0469 μs</td><td>1.9531</td><td>-</td><td>-</td><td>15.63 KB</td>
</tr><tr><td>BulkCopyToSqlServer_cachedTarget_9</td><td>1,194.385 μs</td><td>14.3305 μs</td><td>13.4047 μs</td><td>1.9531</td><td>-</td><td>-</td><td>17.09 KB</td>
</tr><tr><td>BulkCopyToSqlServer_cachedTarget_81</td><td>1,420.294 μs</td><td>13.2123 μs</td><td>12.3588 μs</td><td>3.9063</td><td>-</td><td>-</td><td>31.33 KB</td>
</tr><tr><td>BulkCopyToSqlServer_cachedTarget_729</td><td>3,124.885 μs</td><td>28.9503 μs</td><td>27.0801 μs</td><td>23.4375</td><td>-</td><td>-</td><td>157.48 KB</td>
</tr><tr><td>BulkCopyToSqlServer_uncachedTarget_0</td><td>99.438 μs</td><td>1.4363 μs</td><td>1.3435 μs</td><td>1.5869</td><td>-</td><td>-</td><td>10.45 KB</td>
</tr><tr><td>BulkCopyToSqlServer_uncachedTarget_1</td><td>1,271.974 μs</td><td>13.9532 μs</td><td>13.0518 μs</td><td>1.9531</td><td>-</td><td>-</td><td>23.09 KB</td>
</tr><tr><td>BulkCopyToSqlServer_uncachedTarget_9</td><td>1,301.108 μs</td><td>12.4396 μs</td><td>11.6360 μs</td><td>3.9063</td><td>-</td><td>-</td><td>24.57 KB</td>
</tr><tr><td>BulkCopyToSqlServer_uncachedTarget_81</td><td>1,505.734 μs</td><td>7.4808 μs</td><td>6.9976 μs</td><td>5.8594</td><td>-</td><td>-</td><td>38.81 KB</td>
</tr><tr><td>BulkCopyToSqlServer_uncachedTarget_729</td><td>3,218.563 μs</td><td>24.6221 μs</td><td>23.0315 μs</td><td>23.4375</td><td>-</td><td>-</td><td>164.99 KB</td>
</tr></tbody></table>








This branch (0d1798354aeaf20003f2a0643c69e9355e639fb2): 

<pre><code>
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.508 (2004/?/20H1)
Intel Core i7-8700K CPU 3.70GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.100-rc.1.20452.10
  [Host] : .NET Core 3.1.8 (CoreCLR 4.700.20.41105, CoreFX 4.700.20.41903), X64 RyuJIT
</code></pre>
<pre><code>Job=InProcess  MaxRelativeError=0.002  OutlierMode=DontRemove  
Toolchain=InProcessNoEmitToolchain  IterationCount=15  LaunchCount=1  
MaxIterationCount=5000  WarmupCount=3  
</code></pre>

<table>
<thead><tr><th>                          Method</th><th>    Mean</th><th> Error</th><th>StdDev</th><th>Gen 0</th><th>Gen 1</th><th>Gen 2</th><th>Allocated</th>
</tr>
</thead><tbody><tr><td>BulkInsertTarget_LoadFromTable</td><td>116,304.1 ns</td><td>3,119.14 ns</td><td>2,917.64 ns</td><td>0.9766</td><td>-</td><td>-</td><td>7642 B</td>
</tr><tr><td>BulkCopyToSqlServer_cachedTarget_0</td><td>117.8 ns</td><td>2.39 ns</td><td>2.24 ns</td><td>0.0215</td><td>-</td><td>-</td><td>136 B</td>
</tr><tr><td>BulkCopyToSqlServer_cachedTarget_1</td><td>323,362.4 ns</td><td>17,250.52 ns</td><td>16,136.15 ns</td><td>2.4414</td><td>-</td><td>-</td><td>18074 B</td>
</tr><tr><td>BulkCopyToSqlServer_cachedTarget_9</td><td>1,246,746.4 ns</td><td>24,592.56 ns</td><td>23,003.90 ns</td><td>1.9531</td><td>-</td><td>-</td><td>18096 B</td>
</tr><tr><td>BulkCopyToSqlServer_cachedTarget_81</td><td>1,418,329.2 ns</td><td>15,300.09 ns</td><td>14,311.72 ns</td><td>3.9063</td><td>-</td><td>-</td><td>32676 B</td>
</tr><tr><td>BulkCopyToSqlServer_cachedTarget_729</td><td>3,114,785.2 ns</td><td>12,643.70 ns</td><td>11,826.93 ns</td><td>23.4375</td><td>-</td><td>-</td><td>161853 B</td>
</tr><tr><td>BulkCopyToSqlServer_uncachedTarget_0</td><td>90,180.4 ns</td><td>1,471.91 ns</td><td>1,376.82 ns</td><td>1.2207</td><td>-</td><td>-</td><td>7778 B</td>
</tr><tr><td>BulkCopyToSqlServer_uncachedTarget_1</td><td>401,740.7 ns</td><td>4,448.68 ns</td><td>4,161.29 ns</td><td>3.9063</td><td>-</td><td>-</td><td>25722 B</td>
</tr><tr><td>BulkCopyToSqlServer_uncachedTarget_9</td><td>1,294,510.9 ns</td><td>10,776.88 ns</td><td>10,080.70 ns</td><td>3.9063</td><td>-</td><td>-</td><td>25749 B</td>
</tr><tr><td>BulkCopyToSqlServer_uncachedTarget_81</td><td>1,521,271.2 ns</td><td>5,514.05 ns</td><td>5,157.85 ns</td><td>5.8594</td><td>-</td><td>-</td><td>40332 B</td>
</tr><tr><td>BulkCopyToSqlServer_uncachedTarget_729</td><td>3,227,109.2 ns</td><td>34,976.11 ns</td><td>32,716.67 ns</td><td>23.4375</td><td>-</td><td>-</td><td>169538 B</td>
</tr></tbody></table>