@EamonNerbonne 2020-10-13
Geen haast, wel interessant om ooit te bespreken; het is namelijk aantrekkelijk als we geen uitzonderingen hoeven maken voor inserts van "weinig" data.