@EamonNerbonne 2020-10-16
@fdij @chrarnoldus kunnen jullie hiernaar kijken?  De doelstelling is dat een losse insert nauwelijks nog voordeel bied boven een bulk insert, zodat we luier kunnen zijn (vooral ook in de testen) zonder dat dingen trager worden. Het verschilt toch nog wel een factor drie, dus ik denk dat dat niet helemaal zinloos is, gezien hoe vaak zijn toch enkele records inserten.