@EamonNerbonne 2020-10-16
> Hebben we een test voor meer dan 10 items?

Zo niet moet die er natuurlijk komen.  Maar waar ik me vooral over zorgen maak is over de aanpak - is dit ueberhaupt een goed idee?  Bij welke bulk-insert-options moeten we dit doen?  De oude bulk-insert default had *geen* keep nulls, maar `insert into` gedraagd zich wel zo - moeten we/kunnen we dat zomaar aanpassen (ik zou liefst de scalar en bulk situaties nog even los van deze optimalisatie zo veel mogelijk op elkaar laten lijken, want anders is het zo'n gotcha).