@EamonNerbonne 2020-10-16
@chrarnoldus er waren wel wat tests die toevallig de non-small-batch pad raakten, maar niks expliciets, dus ik heb een extra test toegevoegd.