@chrarnoldus 2020-10-19
> ik zou liefst de scalar en bulk situaties nog even los van deze optimalisatie zo veel mogelijk op elkaar laten lijken, want anders is het zo'n gotcha

lijkt me prima.