@EamonNerbonne 2020-10-19
Ok, ik heb de semantiek van de aanpak zoals jullie bekeken hebben behouden, maar toegevoegd zijn wat tests, en de implementatie is in een losse file gekomen zodat het overzichtelijker is, en ik heb KeepNull optioneel gemaakt, just in case (met test).